Generalise case search mocked pytest fixtures
