+++ unit_tests/caseworker/queues/views/test_cases.py
@@ -38,13 +38,6 @@ def setup(
 
 
 @pytest.fixture
-def mock_cases_search(requests_mock, data_cases_search, queue_pk):
-    encoded_params = parse.urlencode({"page": 1, "flags": []}, doseq=True)
-    url = client._build_absolute_uri(f"/cases/?queue_id={queue_pk}&{encoded_params}")
-    return requests_mock.get(url=url, json=data_cases_search)
-
-
-@pytest.fixture
 def mock_cases_search_team_queue(requests_mock, data_cases_search):
     encoded_params = parse.urlencode({"page": 1, "flags": []}, doseq=True)
     url = client._build_absolute_uri(f"/cases/?queue_id={queue_pk}&{encoded_params}")
@@ -120,12 +113,6 @@ def mock_team_cases_search(requests_mock, data_cases_search):
     data_cases_search["results"]["filters"]["is_system_queue"] = False
     url = client._build_absolute_uri(f"/cases/")
     yield requests_mock.get(url=url, json=data_cases_search)
-
-
-@pytest.fixture
-def mock_cases_search_head(requests_mock):
-    url = client._build_absolute_uri(f"/cases/")
-    yield requests_mock.head(url=re.compile(f"{url}.*"), headers={"resource-count": "350"})
 
 
 def test_cases_home_page_view_context(authorized_client):
